requirements.txt
autopep8==1.6.0
Django==2.2.25
django-filter==2.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
Pillow==9.1.0
psycopg2==2.8.6
pycodestyle==2.8.0
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2

